requirements.txt
celery==5.5.3
redis==4.5.5
flower==1.1.0
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
psycopg2-binary==2.9.6
SQLAlchemy==1.4.47
pytest-asyncio==1.1.0
opentelemetry-api==1.35.0
